htable: avoid undefined shift behavior with large ht->bits
With `-fsanitize=undefined`, I've seen this:

	../ccan/ccan/htable/htable.c:87:17: runtime error: shift exponent 1284229440 is too large for 32-bit type 'int'

Avoid by casting the shifted '1' to size_t prior to the shift.

This large shift exponent was caused by another issue in my code, in
reality that exponent is unreachable (we don't have 1284229440 bit
machines). That said, if we somehow have a 32-bit or larger ht, we'll
run into the same undefined behavior, so let's head it off now.